deployment/backend/backend/requirements.txt
Django==2.2.7
django-cors-headers==3.1.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
mysqlclient==1.4.4
Pillow==6.2.1
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.3
pycryptodome
gunicorn
